LayoutTests/platform/chromium/inspector/console/console-uncaught-exception-in-eval-expected.txt

CONSOLE MESSAGE: line 5: Uncaught Error: Exception in eval.
CONSOLE MESSAGE: line 5: Uncaught Error: Exception in eval.
Tests that when uncaught exception in eval'ed script ending with //@ sourceURL=url is logged into console, its stack trace will have the url as the script source. Bug 47252.

Page reloaded.
Uncaught Error: Exception in eval. evalURL.js:5b evalURL.js:5a evalURL.js:10evalSource evalURL.js:13(anonymous function) evalURL.js:14(anonymous function) console-uncaught-exception-in-eval.html:31
